+++ tests/testthat/test-vegaspec.R
@@ -5,8 +5,8 @@ library("magrittr")
 test_that("as_vegaspec translates", {
 
   # need to go to json and back because of data-frame in vw_ex_mtcars
-  spec_ref <- vw_ex_mtcars %>% as_json() %>% as_vegaspec()
-  spec_json <- as_json(spec_ref)
+  spec_ref <- vw_ex_mtcars %>% vw_as_json() %>% as_vegaspec()
+  spec_json <- vw_as_json(spec_ref)
 
   expect_identical(as_vegaspec(spec_ref), spec_ref)
   expect_identical(as_vegaspec(spec_json), spec_ref)