+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * overhauls the API (#15, with @AliciaSchep)
 
+  * changes `as_json()` -> `vw_as_json()`
   * changes `spec_mtcars` -> `vw_ex_mtcars`
 
 # vegawidget 0.0.2